Creating zod schemas for auth
