+++ src/config/schemas/user/user-dtos.schemas.ts
@@ -1,21 +1,7 @@
 import { z } from 'zod';
 
-import { CheckUserEmailDto, LoginUserDto, RegisterUserDto, UpdateUserDto, UserIdDto } from '../../../domain';
+import { UpdateUserDto, UserIdDto } from '../../../domain';
 import { idSchema } from '../..';
-
-export const registerUserSchema: z.ZodType<RegisterUserDto> = z.object({
-  fullname: z.string().min(5),
-  email: z.string().email(),
-  password: z.string().min(5),
-  avatar: z.string().url(),
-});
-
-export const loginUserSchema: z.ZodType<LoginUserDto> = z.object({
-  email: z.string().email(),
-  password: z.string().min(5),
-});
-
-export const checkUserEmailSchema: z.ZodType<CheckUserEmailDto> = z.object({ email: z.string().email() });
 
 export const userIdSchema: z.ZodType<UserIdDto> = z.object({ user_id: z.lazy(() => idSchema) });
